Include Vuetify in the tests
Very important commit. Mainly because our sandboxed enviroment
needs to be aware of existence of all of those vuetify components
that I'm using. Vuetify documentation seems to be out-of-date because
it encourages use of createLocalVue has been removed from modern
versions of @vue/test-utils. The right way is to include all such
dependencies (that should be plugged to our vue instance), as a value
to the global.plugins array which is the 2nd parameter passed to the
mount function.