Make darkmode toggle based on the props
The darkmode toggle up to this point was pretty much fully DOM dependant.
It held its own state which was the value of the checkbox input.
Now when App is the the only source of truth and I integrated it with
web storage API it's time to make darkmode toggle more reusable
and dependant on the parameters coming from outside sources like
parent component.